refactor: xmodule/course_module.py -> xmodule/course_block.py
