+++ cms/djangoapps/contentstore/views/tests/test_item.py
@@ -28,7 +28,7 @@ from xblock.runtime import DictKeyValueStore, KvsFieldData
 from xblock.test.tools import TestRuntime
 from xblock.validation import ValidationMessage
 from xmodule.capa_block import ProblemBlock
-from xmodule.course_module import DEFAULT_START_DATE
+from xmodule.course_block import DEFAULT_START_DATE
 from xmodule.modulestore import ModuleStoreEnum
 from xmodule.modulestore.django import modulestore
 from xmodule.modulestore.exceptions import ItemNotFoundError
