+++ setup.py
@@ -10,7 +10,7 @@ XBLOCKS = [
     "annotatable = xmodule.annotatable_block:AnnotatableBlock",
     "chapter = xmodule.seq_module:SectionBlock",
     "conditional = xmodule.conditional_block:ConditionalBlock",
-    "course = xmodule.course_module:CourseBlock",
+    "course = xmodule.course_block:CourseBlock",
     "course_info = xmodule.html_module:CourseInfoBlock",
     "customtag = xmodule.template_module:CustomTagBlock",
     "custom_tag_template = xmodule.template_module:CustomTagTemplateBlock",
