+++ xmodule/modulestore/tests/test_split_modulestore.py
@@ -21,7 +21,7 @@ from xblock.fields import Reference, ReferenceList, ReferenceValueDict
 from openedx.core.djangolib.testing.utils import CacheIsolationMixin
 from openedx.core.lib import tempdir
 from openedx.core.lib.tests import attr
-from xmodule.course_module import CourseBlock
+from xmodule.course_block import CourseBlock
 from xmodule.fields import Date, Timedelta
 from xmodule.modulestore import ModuleStoreEnum
 from xmodule.modulestore.edit_info import EditInfoMixin
